backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiosqlite==0.19.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.4
python-dotenv==1.0.0
playwright>=1.40.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
